refactor(refactord header): changed header, created container
change header fron class to functional component, created container
